+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Features
 - Introduce keywords, which are values of type binary, string or atom.
 - Use keywords for member names in object constraints.
+- Add a `values` constraint for strings.
 
 # 1.1.0
 ## Features